requirements.txt
Django==1.11.2
django-crispy-forms==1.6.1
django-recaptcha==1.3.1
Pillow==4.1.1
pytz==2017.2
sorl-thumbnail==12.3
Unidecode==0.4.20
